+++ R/plot_by_pair.R
@@ -15,7 +15,7 @@
 plot_by_pair <-function(df = NULL,
                         bybeat = FALSE,
                         reference = 0,
-                        colourpalette = 'PuOr'){
+                        colourpalette = 'Set2'){
 
   # T. Eerola, Durham University, IEMP project
   # 14/1/2018  
@@ -29,7 +29,9 @@ plot_by_pair <-function(df = NULL,
   m <- tidyr::pivot_longer(data=df$asynch,cols=c(1:dim(df$asynch)[2]),names_to = 'Instrument', values_to = 'ms')
   m$ms <- m$ms*1000
 
-  g1 <- ggplot2::ggplot(m,ggplot2::aes(x=reorder(x=Instrument,-ms,mean),y=ms,fill=Instrument))+
+  m$Instrument <- factor(m$Instrument)
+
+  g1 <- ggplot2::ggplot(m,ggplot2::aes(x=Instrument,y=ms,fill=Instrument))+
     ggplot2::geom_violin(scale = "width",show.legend = FALSE,na.rm=TRUE,alpha=0.7)+
     ggplot2::geom_jitter(height = 0, width = 0.15,size=0.15,show.legend = FALSE,na.rm=TRUE,colour='grey50')+
     ggplot2::stat_summary(fun = "mean",
@@ -48,21 +50,19 @@ plot_by_pair <-function(df = NULL,
   if(bybeat==TRUE){
     #m <- suppressMessages(reshape2::melt(df$asynch,variable.name='Instrument',value.name='ms'))
     m <- tidyr::pivot_longer(data=df$asynch,cols=c(1:dim(df$asynch)[2]),names_to = 'Instrument', values_to = 'ms')
-    
     m$ms<-m$ms*1000
-    
     b <- tidyr::pivot_longer(data=df$beatL,cols=c(1:dim(df$beatL)[2]),names_to = 'variable', values_to = 'value')
-    
     
     m$beatL <- b$value
     m$beatL <- factor(m$beatL)
-
+    m$Instrument <- factor(m$Instrument)
+    
     # Make sure there are sufficient number of colour palette categories for subdivs
     colpal <- RColorBrewer::brewer.pal(4,colourpalette)
     number_of_beats <- length(unique(m$beatL))
     colourpalette <- colorRampPalette(colpal)(number_of_beats)
 
-    g1 <- ggplot2::ggplot(m,ggplot2::aes(x=reorder(x=Instrument,-ms,mean),y=ms,fill=beatL,label=beatL))+
+    g1 <- ggplot2::ggplot(m,ggplot2::aes(x=Instrument,y=ms,fill=beatL,label=beatL))+
       ggplot2::geom_boxplot(outlier.shape = NA,varwidth = F,na.rm=TRUE)+
       ggplot2::geom_hline(yintercept = reference,color='orange4',linetype='dashed')+
       ggplot2::scale_fill_manual(name='Subdiv.',values = colourpalette)+